# src/RedShirt.Adventure.Realm/Models/Characters/Character/CharacterPostRequest.cs
namespace RedShirt.Adventure.Realm.Models.Characters.Character;

public class CharacterPostRequest
{
    public string AccountId { get; init; } = string.Empty;
    public string FirstName { get; init; } = string.Empty;
    public string LastName { get; init; } = string.Empty;
}