# src/RedShirt.Adventure.Realm/Models/Characters/CharacterLocation/CharacterLocationPutRequest.cs
namespace RedShirt.Adventure.Realm.Models.Characters.CharacterLocation;

public class CharacterLocationPutRequest
{
    public int MapId { get; init; }
    public float PositionX { get; init; }
    public float PositionY { get; init; }
    public float PositionZ { get; init; }
    public float Rotation { get; init; }
}