# swagger/RedShirt.Adventure.Realm.Interop/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "RedShirt.Adventure.Realm",
    "version": "1.0.0"
  },
  "paths": {
    "/character/{id}": {
      "delete": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_Patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patchRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterPatchRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDto"
                }
              }
            }
          },
          "304": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/character": {
      "get": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_GetList",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "firstName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "lastName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          },
          {
            "name": "continuationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterSearchResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Character"
        ],
        "operationId": "Character_Post",
        "requestBody": {
          "x-name": "postRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterPostRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDto"
                }
              }
            }
          }
        }
      }
    },
    "/character/{id}/location": {
      "delete": {
        "tags": [
          "CharacterLocation"
        ],
        "operationId": "CharacterLocation_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CharacterLocation"
        ],
        "operationId": "CharacterLocation_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLocationDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "CharacterLocation"
        ],
        "operationId": "CharacterLocation_Patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "patchRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterLocationPatchRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLocationDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "CharacterLocation"
        ],
        "operationId": "CharacterLocation_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "putRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterLocationPutRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLocationDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        }
      }
    },
    "/character/attributes/location": {
      "get": {
        "tags": [
          "CharacterLocation"
        ],
        "operationId": "CharacterLocation_List",
        "parameters": [
          {
            "name": "mapId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "continuationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterLocationSearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CharacterDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "accountId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "CharacterSearchResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "continuationToken": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterDto"
            }
          }
        }
      },
      "CharacterPatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "accountId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CharacterPostRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "accountId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "CharacterLocationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "characterId": {
            "type": "string",
            "format": "guid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "mapId": {
            "type": "integer",
            "format": "int32"
          },
          "positionX": {
            "type": "number",
            "format": "float"
          },
          "positionY": {
            "type": "number",
            "format": "float"
          },
          "positionZ": {
            "type": "number",
            "format": "float"
          },
          "rotation": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "CharacterLocationSearchResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "continuationToken": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterLocationDto"
            }
          }
        }
      },
      "CharacterLocationPatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mapId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "positionX": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "positionY": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "positionZ": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "rotation": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      },
      "CharacterLocationPutRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mapId": {
            "type": "integer",
            "format": "int32"
          },
          "positionX": {
            "type": "number",
            "format": "float"
          },
          "positionY": {
            "type": "number",
            "format": "float"
          },
          "positionZ": {
            "type": "number",
            "format": "float"
          },
          "rotation": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
    }
  }
}